Refactor/dotnet 8 (#42)
* Refactor code and update to .net 8

* Update dockerfile dotnet version

+++ PetStore.API/Services/ExternalServices/StripeService.cs
@@ -1,7 +1,4 @@
 ﻿using Stripe;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace PetStore.API.Services.ExternalServices
@@ -14,7 +11,7 @@ namespace PetStore.API.Services.ExternalServices
             {
                 Source = tokenId,
                 Currency = "USD",
-                Amount = (long)amount*100
+                Amount = amount * 100
             };
             var service = new ChargeService();
             Charge charge = service.Create(options);
@@ -23,7 +20,7 @@ namespace PetStore.API.Services.ExternalServices
 
         public async Task<string> GetStatusAsync(string chargeId)
         {
-            var service = new ChargeService(); 
+            var service = new ChargeService();
             return (await service.GetAsync(chargeId)).Status;
         }
     }
